build: enable -Wall -Wextra -Wpedantic project-wide and clear all warnings
The tree compiled at the compiler's default warning level; a warning
dry run over all 154 non-test and 45 test sources surfaced 65+9
warnings across 24 files, all fixed here:

- unused params marked [[maybe_unused]] (44) or left unnamed in lambdas,
  keeping the interface-documenting names
- sign-compare fixed via explicit casts where the semantics were already
  unsigned (metrics histogram, edit-history size cap, presence status
  index)
- RedisAuthStore::Impl init list reordered to match member declaration
  order
- dead locals removed (two unused 'now' snapshots, an unused test
  helper)
- FeatureHandlers aggregate init switched to designated form with the
  later-assigned npc_service_id explicit
- greedy hex escapes in WebSocket frame tests split into adjacent
  literals ("\x82\x05ab" parsed \x05ab as one out-of-range escape,
  so the tests did not feed the bytes their comments describe)
- dangling-else around gtest macros braced; GCC 15's -Wfree-nonheap-object
  false positive on vector<uint8_t>::push_back avoided by building the
  length header as std::array

Warnings are on by default (CHIRP_ENABLE_WARNINGS, GNU/Clang); CI build
and smoke jobs pass -DCHIRP_WARNINGS_AS_ERRORS=ON to keep the tree clean.
KNOWN_UNCOVERABLE line numbers realigned for the shifted sources.

+++ libs/common/include/common/metrics.h
@@ -99,7 +99,7 @@ public:
       sum += value;
       // Find bucket
       for (size_t i = 0; i < kNumBuckets; ++i) {
-        if (value <= buckets[i].load()) {
+        if (static_cast<uint64_t>(value) <= buckets[i].load()) {
           ++buckets[i];
           break;
         }
